sw_services: xilplmi: MISRA C Required Violation Rule 10.4 Fixes
MISRA C Rule 10.4 States: Both operands of an operator in which the usual
arithmetic conversions are performed shall have the same essential type category

This patch fixes MISRA C Rule 10.4 Violation by using same type variables in
logical operations


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_debug.c
@@ -84,7 +84,7 @@ int XPlmi_InitUart(void)
 	XUartPsv UartPsvIns;
 	XUartPsv_Config *Config;
 
-	for (Index = 0U; Index < XPAR_XUARTPSV_NUM_INSTANCES; Index++) {
+	for (Index = 0U; Index < (u8)XPAR_XUARTPSV_NUM_INSTANCES; Index++) {
 
 		Status = XPlmi_MemSetBytes(&UartPsvIns, sizeof(XUartPsv),
 				0U, sizeof(XUartPsv));
